requirements.txt
git+https://github.com/Pycord-Development/pycord.git
python-dotenv==0.19.0
buttons
asyncpraw
requests
asyncio